Separation of Google Play Services Google Maps logic from GeoPackage functionality
