+++ geopackage-sdk/build.gradle
@@ -4,7 +4,7 @@ apply plugin: 'signing'
 
 group = "mil.nga.geopackage"
 archivesBaseName = "geopackage-android"
-version = "1.1.2"
+version = "1.2.0"
 def remotePublish = false
 
 android {
